requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
astrapy==2.0.1
asttokens==2.4.1
attrs==25.3.0
autograd==1.8.0
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
backoff==2.2.1
banks==2.1.2
bcrypt==4.3.0
beartype==0.18.5
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
bs4==0.0.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
environs==11.2.1
executing==2.2.0
filetype==1.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
fnllm==0.3.0
fonttools==4.58.2
frozenlist==1.6.0
fsspec==2024.12.0
future==1.0.0
gensim==4.3.3
graph-retriever==0.8.0
graphrag==2.3.0
graspologic==3.4.1
graspologic-native==1.2.5
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
hyppo==0.4.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
lancedb==0.17.0
langchain==0.3.25
langchain-astradb==0.6.0
langchain-community==0.3.24
langchain-core==0.3.64
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-graph-retriever==0.8.0
langchain-neo4j==0.4.0
langchain-openai==0.3.21
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.69
langsmith==0.3.45
language_data==1.3.0
llama-cloud==0.1.23
llama-cloud-services==0.6.30
llama-index==0.12.40
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.3
llama-index-core==0.12.40
llama-index-embeddings-openai==0.3.1
llama-index-graph-stores-neo4j==0.4.6
llama-index-indices-managed-llama-cloud==0.7.4
llama-index-llms-openai==0.4.3
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.30
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
murmurhash==1.0.13
mypy==1.15.0
mypy_extensions==1.1.0
neo4j==5.28.1
neo4j-graphrag==1.7.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.84.0
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pdfminer.six==20250506
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
POT==0.9.5
preshed==3.0.10
propcache==0.3.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.20.0
pymongo==4.13.0
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.6.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scikit-learn==1.7.0
scipy==1.12.0
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
statsmodels==0.14.4
striprtf==0.0.26
tenacity==9.1.2
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
toml==0.10.2
tqdm==4.67.1
typer==0.15.4
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
uuid6==2024.7.10
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
